Modified code structure to implement level-2 asa_wvw_2p reader
